requirements.txt
torch
torchvision
gradio
pillow
numpy
tqdm